2023-07-08 userData API
- 로그인 시 유저 정보가 슬라이스에 담김
- data.json 파일에서 가져온 데이터를 data API에서 추출한 데이터로 변경
